Process WM_IME_SETCONTEXT correctly
"If the application draws the composition window, the default IME window does not have to show its composition window. In this case, the application must clear the ISC_SHOWUICOMPOSITIONWINDOW value from the lParam parameter before passing the message to DefWindowProc or ImmIsUIMessage. To display a certain user interface window, an application should remove the corresponding value so that the IME will not display it."

See https://docs.microsoft.com/en-us/windows/win32/intl/wm-ime-setcontext