Fix incorrect result check for wcscpy_s and wcsnlen
"If dest or src is a null pointer, or if the destination string size dest_size is too small, the invalid parameter handler is invoked, as described in Parameter Validation. If execution is allowed to continue, these functions return errno value".

So result is not HRESULT, but errno_t.

See https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/strcpy-s-wcscpy-s-mbscpy-s